Changed PageExpiredException to GoneException
